Fixing Roulette imports that broke with components reorganization
